move log-request to middlewares
